python-value-pairs: Use _py_string_from_string()
Previously, we used `PyBytes_FromStringAndSize()`, and it was
only tested with `python2`.

Our example `python()` destination has this `send()`:
```
def send(self, msg):
    with open('test-python.log', 'a') as f:
        f.write('{DATE} {HOST} {MSGHDR}{MSG}\n'.format(**msg))
```

In `python2` it does not matter, if we fill the `msg` dict with
`bytes`-like data (which is generated with
`PyBytes_FromStringAndSize()`), the `test-python.log` file will be
written in `text` mode insead of `byte` mode, because it is the default.

However, in `python3` when we open the file in the default `text` mode,
and we write `bytes`-like object in it, the output will still be in
`byte` mode.

This causes a different behavior, which is fixed by storing the data in
`text` mode, with `_py_string_from_string()`.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>