python: fixing printing None instead of the actual exception
If an exception is triggered in python code, such as with the following
configuration[1], normally a message reports in which function what kind of
exception happened. If syslog-ng was compiled with recent clang (10.0) instead of
showing the exception the default None was displayed.

I did not dig into deep enough why this only happens with clang and not with gcc,
it is a good idea - just like in case of ERRNO - to save the exception before
anything else, this patch does exactly that.

Note: also unit test fails due to this issue with clang

[1] configuration that could trigger this issue
```
python {

from syslogng import LogSource

class S(LogSource):
  def init(self, options):
    raise ValueError('exception')

  def run(self):
    pass

  def request_exit(self):
    pass

};

log {
  source {
    python(
        class(S)
    );
  };
```

Signed-off-by: Kokan <kokaipeter@gmail.com>