dbld: run setup before prepare-release
Our docker arguments add volumes to the container. If the referenced
directories are not present, it creates them with restricted permissions.

The `release` job tries to create a dir inside one of those dirs, and it
fails with insufficient permissions.

We should create the directories ourselves, and we have a target for
that.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>