github-actions: add "Version bump" job
This job creates a version bump branch based on master, and opens a PR
from it.

The branch will have the version bumped in all necessary parts of the
code, and it will generate the NEWS.md content from the NEWS snippets.

Organization members have the chance to add or modify the commits on
the branch, if necessary.

The job will also generate tarballs, and reference them in the PR's
description.

The job cannot be started, if there is an open PR with the version
given.

Manual steps:
 1. Navigate to https://github.com/syslog-ng/syslog-ng/actions.
 2. Choose the "Version bump" job.
 3. Click "Run workflow".
 4. Fill the new release version, and click "Run workflow".
 5. When the PR is opened, double check the changes, if necessary,
    add commits.
 6. When the PR is merged, the "Draft release" job will automatically
    start.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>