github-actions: add "Draft release" job
This job creates a draft release into GitHub's releases section.

It is automatically started, when a "Release candidate" PR is merged.
Its progress is commented at the "Release candidate" PR.

The job can also be started manually, if necessary.

When the job is finished, one must visit
https://github.com/syslog-ng/syslog-ng/releases, double check the
generated draft release, and manually release it.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>