config_option_database: refactor file handling in yacc2xml()
_create_temp_file_with_content() creates a file with the given
content.

As we use NamedTemporaryFile, when the python object is freed,
the file itself if deleted.

We don't need to seek to the beginning of the file, as we only pass its
path to the bison binary.

xml_filepath is the output file path of bison. We don't need to come up
with an unused file path in such a complex way, it is easier to use
the temp_yacc_file with a suffix.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>