qdisk: factor out _calculate_new_read_head_position()
This commit changes the atomicity of changing the value of read_head.

It is important, because read_head is on an mmap-ed region, and if a
crash occurs, we still want to have a valid disk-buffer.

The old could change read_head twice, if the first change pointed to
a position, which nedded wrapping, it was then changed again to the
beginning of the file (QDISK_RESERVED_SPACE).
In this case, if a crash occurred between the two, we did not wrap the
read_head, so in the next pop_head after the restart, we might have
read from an invalid space.

The new code calculates the new read_head position with a local variable
and sets read_head once to the valid position. So we can either crash
before setting or after setting. If we crash after we already changed
the read_head, so everything is fine, if we crashed before, we will try
to read the same message again after the restart. Either case we read
from a valid position.

In conclusion the new logic is safer in case of crash.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>