logqueue-disk: serialize msg outside log_queue_disk_write_message()
We are planning to move the serialization outside of the LogQueue lock.
Currently we are forced to serialize after we acquired the lock,
because log_queue_disk_write_message() is called with the lock held.

If we make log_queue_disk_write_message() receive the serialized message
instead of the LogMessage object, we have more control of when we do
the serialization.

This commit moves the serialization to LogQueueDisk::_push_tail().
It has an unfortunate effect, that we sometimes serialize the message,
even if we will not use that (in some cases of nonreliable disk-buffer).
This will be fixed in a later commit.

In this commit we do not yet move the serialization outside the lock,
we will have to investigate 2 places to do that:
 * LogQueueDisk::_push_tail()
 * LogQueueDiskNonReliable::_move_messages_from_overflow()

Doing so will be done in the following commits.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>