logqueue-disk: serialize outside of the lock in _push_tail()
It is common to have multiple source threads.

When we use disk-buffer, the message's serialization and writing it to
the disk happen in the source thread.

We only need to acquire the lock when we are operating on QDisk.
As the serialization happens before sending it to the QDisk, we can
easily move the serialization outside the lock.

This can yield significant performance improvement, because while the
source threads are doing their slow serialization, the destination
thread can process messages from the QDisk.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>