logqueue-disk: add virtual function: need_serialize_hint
In some cases (sometimes in nonreliable), we do not need to serialize
the message, when pushing it to the queue.

This new virtual function lets the LogQueueDisk descendant decide
whether it needs serialized message.

This is still a hint, because we will release the lock after we decided
if we need serialized message or not, then we will lock again.
Between the two locks the need for serialized message might have
changed, so each LogQueueDisk descendant MUST adjust to it.

In this commit we assume that we need serialized message every time,
we will add the actual implementation in the next commit.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>