stats-cluster: add functions to clone StatsClusterGroupInit
StatsClusterGroupInit is part of StatsClusterKey.

We have a function, which clones StatsClusterKey, but it does not
deep-copies its StatsClusterGroupInit.

This is not a problem when we only have string literals in
StatsClusterGroupInit::name/names, but after the previous commit
we can have heap-allocated strings there.

This commit adds the necessary virtual functions and function calls
to make it possible for a StatsClusterGroupInit to implement its own
clone() and cloned_free() functions.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>