package-indexer: make --run-id mandatory for the "nightly" suite
We are planning to run the package indexer in GitHub Actions,
automatized. Previously, when we indexed the nightly packages all of the
available packages were processed in the incoming/nightly directory.
It was a workaround because we could not communicate between the
package generating job in GitHub Actions and our internal indexing job,
so we could not tell the indexing job which dynamically created
directory we want to index. We are now able to do that.

This, and the next couple of commits will do that.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>