github-actions: add stable-release workflow
The draft-release workflow ends with:
  1. packages created and uploaded to Azure/incoming
  2. a newly opened draft release on GitHub

At this point we still can go back and revert any action, as no new
package has been indexed and no new tag has been pushed.

After we made sure, that everything is working as expected, we can
finalize the draft release, then the indexing can start.

Ideally we would do a "test" indexing and "test" production docker image
generation before creating the draft-release, so we can be sure, that
they will succeed after finalizing the draft release, but this will only
be done in a later PR.

The newly added stable-release workflow hooks to the draft release
finalization by checking the tag pushed by it. At this point we only
index and test the packages, but in the next couple commits we will move
the docker image publication to this workflow, too.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>