diskq-global-metrics: get rid of watch_dir()
Unfortunately, it was not the best way to go.

At the point when it was called it watched the dir which was configured.
However, changing the dir between reloads is allowed, but has no effect,
the configured dir will not be used. This means the code watched a dir
unnecessarily and did not watch a dir which should have been watched.

This means that we can only know which dir to watch, in acquire_queue()
where we have the persist entry and we know the dir. At that point we
already call file_acquired(), so there is no need to call watch_dir()
before it, the former can do everything.

It will also come in handy for a microoptimization of the abandoned
queue file metrics.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>