syslog-format: move syslog related tags to the core
To ensure that syslog specific formatting issues can be reported efficiently
by tags, register those as predefined tags in the core.

I made an attempt to register these to be syslogformat specific, however
parts of the configuration may start using the new tags _before_
syslogformat is loaded, causing an incorrect registration of the tags in
question.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>