filterx: rework setattr() and set_subscript() virtual fns
When we will write filterx based parsers, these functions will
be used extensively. A recurring pattern is creating a dict or
list, storing it as a kv pair in a parent dict with set_subscript()
then wanting to add elements to it.

The current interface puts the responsibility to the Object's
implementation to clone the value in the setter functions, and it
always happens there.

For our internal parser codes, these clone() calls just waste
resources, and makes it awkward to set the mutable value, and get
the stored value immediately so it can modify it.

I have tried 3-4 different ways to solve this, but eventually came
to this solution, which is the simplest and most performant:
The Object interface's now requires the implementation to accept
the new value, and modify it, so that it represents the stored
value. No cloning must happen there, we only clone on the end user
related parts (expr-set-subscript and expr-setattr).

Our parser codes can simply set a mutable value, and operate on
the same FilterXObject immediately, but must take care of cloning,
where it is necessary (however it is extremely rare).

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>