filterx: add list and dict interfaces
We do not want to have generic lists and dicts in filterx, as
converting to and from that intermediate layer wastes resources.

Instead we will have an interface for them, and we will write our
parser codes by using that interface. We will also have multiple
implementations (json and otel). After all these we just have to
connect the two somehow.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>